Read in polylink data.

Reads in data. Optionally, can merge gene sets that share more than a specified proportion of genes, or remove gene sets with less/more than a specified number of genes.

ReadSetObjTables(
  in.path = "./",
  population = NA,
  minsetsize = 10,
  maxsetsize = 1000,
  merge.set.prop = 1,
  n.cores = NA
)

Arguments

in.path
character: pathway to input files

population
character: label used to identify input files. Three input files are required, specified by 'setInfo', 'setObj', and 'objInfo'.

minsetsize
integer: minimum number of genes required in gene set

maxsetsize
integer: maximum number of genes required in gene set

n.cores
integer: number of cores to run in parallel

min.sim
double: sets sharing at least this proportion of genes will be merged

Value

Returns a list with three elements.