polylinkR: Polygenic Pathway Enrichment with Linkage Structure

Perform gene-based pathway enrichment analysis while accounting for linkage disequilibrium (LD) amongst loci. Uses a permutation approach that preserves linkage structure when generating null distributions of pathway scores. This reduces false positives compared to standard permutation methods. This version is under active development and must not be used as the sole basis for production or clinical decisions without independent validation.

See also

Author

Maintainer: PolyLinkR authors polylinkr@acad.edu.au

Other contributors:

ACAD University of Adelaide [copyright holder, funder]